Removed everything that is unsafe.Pointer to use uintptr. Any thing that requires a unsafe.Pointer to remain will have to use a refcounter.

+++ defaultengine.go
@@ -19,7 +19,7 @@ func (e StdEng) makeArray(arr *array, t Dtype, size int) {
 	arr.t = t
 	arr.L = size
 	arr.C = size
-	arr.Ptr = unsafe.Pointer(&s[0])
+	arr.Ptr = uintptr(unsafe.Pointer(&s[0]))
 	arr.fix()
 }
